+++ cBugId.py
@@ -32,22 +32,35 @@ class cBugId(object):
     rImportantStdOutLines = None,
     rImportantStdErrLines = None,
     bGenerateReportHTML = False,
+    fFailedToDebugApplicationCallback = None,
     fApplicationRunningCallback = None,
-    fExceptionDetectedCallback = None,
-    fApplicationExitCallback = None,
-    fFinishedCallback = None,
+    fApplicationSuspendedCallback = None,
+    fApplicationResumedCallback = None,
+    fMainProcessTerminatedCallback = None,
     fInternalExceptionCallback = None,
+    fFinishedCallback = None,
   ):
     # Replace fFinishedCallback with a wrapper that signals the finished event.
     # This event is used by the fWait function to wait for the process to
     # finish.
-    oBugId.__fExternalFinishedCallback = fFinishedCallback;
-    oBugId.__fExternalInternalExceptionCallback = fInternalExceptionCallback;
+    oBugId.__fFailedToDebugApplicationCallback = fFailedToDebugApplicationCallback;
+    oBugId.__fApplicationRunningCallback = fApplicationRunningCallback;
+    oBugId.__fApplicationSuspendedCallback = fApplicationSuspendedCallback;
+    oBugId.__fApplicationResumedCallback = fApplicationResumedCallback;
+    oBugId.__fMainProcessTerminatedCallback = fMainProcessTerminatedCallback;
+    oBugId.__fInternalExceptionCallback = fInternalExceptionCallback;
+    oBugId.__fFinishedCallback = fFinishedCallback;
+    
     oBugId.__oFinishedEvent = threading.Event();
     oBugId.__bStarted = False;
+    # If debugging fails, this is set to a string that describes the reason why:
+    oBugId.sFailedToDebugApplicationErrorMessage = None;
+    # If an internal exception happens, this is set to the Exception object:
+    oBugId.oInternalException = None;
+    # If a bug was found, this is set to the bug report:
+    oBugId.oBugReport = None;
     # Run the application in a debugger and catch exceptions.
     oBugId.__oCdbWrapper = cCdbWrapper(
-      oBugId = oBugId,
       sCdbISA = sCdbISA,
       asApplicationCommandLine = asApplicationCommandLine,
       auApplicationProcessIds = auApplicationProcessIds,
@@ -58,11 +71,13 @@ class cBugId(object):
       rImportantStdOutLines = rImportantStdOutLines,
       rImportantStdErrLines = rImportantStdErrLines,
       bGenerateReportHTML = bGenerateReportHTML,
-      fApplicationRunningCallback = fApplicationRunningCallback,
-      fExceptionDetectedCallback = fExceptionDetectedCallback,
-      fApplicationExitCallback = fApplicationExitCallback,
-      fFinishedCallback = oBugId.__fFinishedHandler,
+      fFailedToDebugApplicationCallback = oBugId.__fFailedToDebugApplicationHandler,
+      fApplicationRunningCallback = oBugId.__fApplicationRunningHandler,
+      fApplicationSuspendedCallback = oBugId.__fApplicationSuspendedHandler,
+      fApplicationResumedCallback = oBugId.__fApplicationResumedHandler,
+      fMainProcessTerminatedCallback = oBugId.__fMainProcessTerminatedHandler,
       fInternalExceptionCallback = oBugId.__fInternalExceptionHandler,
+      fFinishedCallback = oBugId.__fFinishedHandler,
     );
   
   def fStart(oBugId):
@@ -109,11 +124,44 @@ class cBugId(object):
         "You must call cBugId.fStart() before calling cBugId.fbFinished()";
     oBugId.__oFinishedEvent.isSet();
   
-  def __fFinishedHandler(oBugId, oBugId2, oBugReport):
+  # Wrap all callbacks to provide this cBugId instance as the first argument.
+  def __fFailedToDebugApplicationHandler(oBugId, sErrorMessage):
+    # Save error message if application cannot be debugged.
+    oBugId.sFailedToDebugApplicationErrorMessage = sErrorMessage;
+    oBugId.__oFinishedEvent.set();
+    if oBugId.__fFailedToDebugApplicationCallback:
+      oBugId.__fFailedToDebugApplicationCallback(oBugId, sErrorMessage);
+  
+  def  __fApplicationRunningHandler(oBugId):
+    if oBugId.__fApplicationRunningCallback:
+      oBugId.__fApplicationRunningCallback(oBugId);
+  
+  def  __fApplicationSuspendedHandler(oBugId):
+    if oBugId.__fApplicationSuspendedCallback:
+      oBugId.__fApplicationSuspendedCallback(oBugId);
+  
+  def  __fApplicationResumedHandler(oBugId):
+    if oBugId.__fApplicationResumedCallback:
+      oBugId.__fApplicationResumedCallback(oBugId);
+  
+  def __fExceptionDetectedHandler(oBugId):
+    if oBugId.__fExceptionDetectedCallback:
+      oBugId.__fExceptionDetectedCallback(oBugId);
+  
+  def __fMainProcessTerminatedHandler(oBugId):
+    if oBugId.__fMainProcessTerminatedCallback:
+      oBugId.__fMainProcessTerminatedCallback(oBugId);
+  
+  def __fInternalExceptionHandler(oBugId, oException):
+    # Save internal exception.
+    oBugId.oInternalException = oException;
+    if oBugId.__fInternExceptionCallback:
+      oBugId.__fInternExceptionCallback(oBugId, oException);
+  
+  def __fFinishedHandler(oBugId, oBugReport):
+    # Save bug report, if any.
     oBugId.oBugReport = oBugReport;
     oBugId.__oFinishedEvent.set();
-    oBugId.__fExternalFinishedCallback and oBugId.__fExternalFinishedCallback(oBugId2, oBugReport);
+    if oBugId.__fFinishedCallback:
+      oBugId.__fFinishedCallback(oBugId, oBugReport);
 
-  def __fInternalExceptionHandler(oBugId, oBugId2, oException):
-    oBugId.oInternalException = oException;
-    oBugId.__fExternalInternExceptionCallback and oBugId.__fExternalInternExceptionCallback(oBugId2, oException);